+++ xhttp/retry_test.go
@@ -72,6 +72,36 @@ func testRetryTransactorNoRetries(t *testing.T) {
 	assert.True(transactorCalled)
 }
 
+func testRetryTransactorStatus(t *testing.T) {
+	var (
+		assert  = assert.New(t)
+		require = require.New(t)
+
+		transactorCount = 0
+		statusCheck     = 0
+		transactor      = func(*http.Request) (*http.Response, error) {
+			response := http.Response{
+				StatusCode: 429 + transactorCount,
+			}
+			transactorCount++
+			return &response, nil
+		}
+
+		retry = RetryTransactor(RetryOptions{
+			Retries: 5,
+			ShouldRetryStatus: func(status int) bool {
+				statusCheck++
+				return status == 429
+			},
+		}, transactor)
+	)
+
+	require.NotNil(retry)
+	retry(httptest.NewRequest("GET", "/", nil))
+	assert.Equal(2, transactorCount)
+	assert.Equal(2, statusCheck)
+}
+
 func testRetryTransactorAllRetriesFail(t *testing.T, expectedInterval, configuredInterval time.Duration, retryCount int) {
 	var (
 		assert          = assert.New(t)
@@ -79,9 +109,15 @@ func testRetryTransactorAllRetriesFail(t *testing.T, expectedInterval, configure
 		expectedRequest = httptest.NewRequest("GET", "/", nil)
 		expectedError   = &net.DNSError{IsTemporary: true}
 		counter         = generic.NewCounter("test")
+		urls            = map[string]int{}
 
 		transactorCount = 0
 		transactor      = func(actualRequest *http.Request) (*http.Response, error) {
+			if _, ok := urls[actualRequest.URL.Path]; ok {
+				urls[actualRequest.URL.Path]++
+			} else {
+				urls[actualRequest.URL.Path] = 1
+			}
 			transactorCount++
 			assert.True(expectedRequest == actualRequest)
 			return nil, expectedError
@@ -98,6 +134,11 @@ func testRetryTransactorAllRetriesFail(t *testing.T, expectedInterval, configure
 					slept++
 					assert.Equal(expectedInterval, actualInterval)
 				},
+				UpdateRequest: func(request *http.Request) {
+					if _, ok := urls[request.URL.Path]; ok {
+						request.URL.Path += "a"
+					}
+				},
 			},
 			transactor,
 		)
@@ -110,6 +151,9 @@ func testRetryTransactorAllRetriesFail(t *testing.T, expectedInterval, configure
 	assert.Equal(1+retryCount, transactorCount)
 	assert.Equal(float64(retryCount), counter.Value())
 	assert.Equal(retryCount, slept)
+	for _, v := range urls {
+		assert.Equal(1, v)
+	}
 }
 
 func testRetryTransactorFirstSucceeds(t *testing.T, retryCount int) {
@@ -227,4 +271,5 @@ func TestRetryTransactor(t *testing.T) {
 
 	t.Run("NotRewindable", testRetryTransactorNotRewindable)
 	t.Run("RewindError", testRetryTransactorRewindError)
+	t.Run("StatusRetry", testRetryTransactorStatus)
 }